// interface.ts
export interface ordStatusDetail_ {
    I_ORDER: string;
    C_STATUS: string;
    I_DETAIL: number;
    C_PROCESSING: string;
    C_ROUND_TYPE: string;
    D_CREATE: string;
    D_CHANGE: string;
    D_TIMESTAMP: string;
    C_CANCEL_REASON	: string;
    C_SEND_EDN : string;
    D_SEND_EDN: string;
    Update :  string;
    Delete: string;
  }